requirements.txt
py-cord[voice]==2.4.1
yt-dlp==2023.11.16
spotipy==2.23.0
python-dotenv==1.0.0
aiohttp==3.8.6